==> analysis/80_exploratory_analysis.Rmd <==
---
title: "Exploratory analyses"
author: "Petra Schwalie"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    toc: true
    toc_float: true
    code_folding: hide
params:
  data_collection: "UUID"
  output_collection: "UUID"
  .arv_save:
    value:
      collection: "UUID"
      filename: "{basename(html_filename)}"
---


```{r libs, message=FALSE, warning=FALSE}
library(Biobase)
library(assertthat)
library(biokitr)
#library(tidyverse)
library(RColorBrewer )
 library(ComplexHeatmap)
library(umap)
library(ggpubr)

requireNamespace("BioQC")
# require minimal BioQC version
stopifnot(packageVersion("BioQC") >= package_version("1.21.2"))

#### Function to scale the rows for plotting
scaleRows=function (x) 
{
    m = apply(x, 1, mean, na.rm = T)
    s = apply(x, 1, sd, na.rm = T)
    return((x - m)/s)
}

# -----------------------------------------------------------------------------
# Colors

### Initialise colors
redblue<-colorRampPalette(c("snow3","orange","darkred"))
redblue2<-colorRampPalette(c("blue","snow3","darkred"))

```


```{r random seed}
set.seed(42)
```

```{r}
set.seed(42)

addline_format <- function(x,...){
    gsub('\\s','\n',x)
}
```

Load ExpressionSet object.
```{r load_eset}
  eset <- aws.s3::s3readRDS("tpm_eset.Rds", bucket = params$output_collection)
  eset_counts <- aws.s3::s3readRDS("counts_eset.Rds", bucket = params$output_collection)
  
  #check which samples are gred
  #samples_annot <- aws.s3::s3read_using(FUN = read.csv,object ="full_samples_anno.tsv", bucket = params$data_collection, sep="\t")
  #distinct(samples_annot[,c("SOURCE","STUDYID")])
  
  # Remove gRED samples from exploratory analysis 
  eset=eset[,(which(!pData(eset)$STUDYID%in%c("GO29293", "GO28915" ,"GO29294")))]
  eset_counts=eset_counts[,(which(!pData(eset_counts)$STUDYID%in%c("GO29293", "GO28915" ,"GO29294")))]
  
  ### Prepping ematrix and annotation for
  #TPM = Biobase::exprs(eset)
  pseudo=0.01
  ematrix=data.matrix(log2(exprs(eset)+pseudo))

  
  Design = Biobase::pData(eset) #sampleAnnot_complete.dt 
  Annot = Biobase::fData(eset)
  rownames(ematrix)=Annot$symbol
  
  # cleaning information 
  # UROTHELIAL BLADDER CANCER	 == gRED
  # BLADDER = pRED 
  
  # NSCLC = gRED
  # LUNG = pRED
  
  ### Add correct indication annotation 
  Design_HA=read.table("../data/Design_indications_final.tsv",sep="\t")
  Design[Design_HA$A_SampleID,"INDICAT"]=Design_HA$HA_Indication

  Design$BMFLOC [ grep("abdominal_cavity", Design$BMFLOC) ]="gastrointestinal_tract"
  Design$BMFLOC [ grep("esophagus", Design$BMFLOC) ]="gastrointestinal_tract"
  Design$BMFLOC [ grep("intestines", Design$BMFLOC) ]="gastrointestinal_tract"
  Design$BMFLOC [ grep("stomach", Design$BMFLOC) ]="gastrointestinal_tract"
  #Design$BMFLOC [ grep("other", Design$BMFLOC) ]="other or unknown"
  Design$BMFLOC [ grep("bronchus", Design$BMFLOC) ]="lung"
  
  over5_i = Design %>% group_by( INDICAT) %>% tally() %>% filter (n > 5) %>% tidyr::drop_na(INDICAT)
  over5_l = Design %>% group_by( BMFLOC) %>% tally() %>% filter (n > 5) %>% tidyr::drop_na(BMFLOC)
  
  Design$INDICAT [which(!Design$INDICAT%in%over5_i$INDICAT)] <- "OTHER"
  Design$BMFLOC [which(!Design$BMFLOC%in%over5_l$BMFLOC)] <- "other"

  Design$INDICAT=substr(make.names(Design$INDICAT),1,14)
  Design$BMFLOC=substr(make.names(Design$BMFLOC),1,14)
  #Design$strat=paste(substr(make.names(Design$INDICAT),1,10),substr(make.names(Design$BMFLOC),1,10),sep="_")
  
  Biobase::pData(eset)=Design
  
  Annotk=cbind(rownames(Annot),Annot$symbol, Annot$size, Annot$type, Annot$desc)
  colnames(Annotk)=c("GeneID","GeneSymbol","GeneLength","GeneType","GeneDesc")
  Annotk=as.data.frame(Annotk)
  Annotk[,"GeneLength"]=as.numeric(Annotk[,"GeneLength"])
  
  # DGEList including all genes
  dge.allGenes <- edgeR::DGEList(counts=exprs(eset_counts), samples=Design, genes=Annotk)
 
  # DGEList after removing lowly expressed genes
  num_samples_expressed_cutoff=-1
  cpm_cutoff=1
  dge <- removeLowExpressedGenes(dge.allGenes, numSamplesExpressedCutoff = num_samples_expressed_cutoff,
                               cpmCutoff = cpm_cutoff, groupColumn = "BMFLOC")
  
  cpms <- getCpm(dge, log = FALSE)
  numExpressed <- colSums(cpms > 1)
  dge$samples$N_genes_CPM_1 <- numExpressed
  
  dgelog=dge
  dgelog$counts=data.matrix(log2(exprs(eset)+pseudo))[rownames(dge$counts),]
  
  # Define custom colors (color_vars for which no entry is added use default ggplot colors)
  # The variables used for colors
  colorVars=c("INDICAT", "CD8IMMPH", "BMFLOC","STUDYID","STATUS")

```


Load Signatures object.
```{r load_sig_set}
sig_scores <- aws.s3::s3readRDS("sig_scores.Rds", bucket = params$output_collection)
sig_scores <- sig_scores[,rownames(Design)]
```



```{r load functions}
source("../src/eset_transformation.R")
```

# Signature Analysis

## Signature Heatmap

```{r heatmap.full, warning=FALSE, comment=FALSE, fig.height = 5, fig.width = 14, fig.align = "center"}


choicesiglarge=c("sc_Bcell" ,"sc_Fibroblast" , "sc_Endothelial" ,"sc_Epithelial", "sc_Granulocyte",  "sc_MelMelanoma"  ,  "sc_Myeloid", 
              "sc_Plasma" , "sc_CD4Tcell"  ,  "sc_CD8Tcell" , "sc_RegTcell"   , "sc_cDC1"      ,  "sc_cDC2"  ,"sc_pDC" ,"sc_Hematopoietic" , "sc_Blymphocyte"  ,
              "sc_Proliferation"  ,"sc_Macrophage_MARCO" , "sc_Cytotoxicity"  ,"sc_ExhCD8Tcell"  , "sc_Macrophage_MSR1"  , "sc_Macrophage_CXCL9","sc_Monocyte"  ,
              "sc_Macrophage" ,  "sc_Hepatocyte"       , "sc_Adipocyte"  , "sc_InflamFibroblast"  ,"sc_Mast"  ,"sc_ColorectalCancer", "hallmark_HALLMARK_TGF_BETA_SIGNALING",
               "hallmark_HALLMARK_INTERFERON_ALPHA_RESPONSE","hallmark_HALLMARK_INTERFERON_GAMMA_RESPONSE", "hallmark_HALLMARK_MYC_TARGETS_V1" ,"hallmark_HALLMARK_OXIDATIVE_PHOSPHORYLATION"   ,
              "hallmark_HALLMARK_INFLAMMATORY_RESPONSE" ,"hallmark_HALLMARK_ANGIOGENESIS"  , "hallmark_HALLMARK_MYC_TARGETS_V2"   , "hallmark_HALLMARK_APOPTOSIS"          )

choicesig=c("sc_Fibroblast" , "sc_Endothelial" , "sc_Myeloid", 
               "sc_RegTcell"   , "sc_Hematopoietic" , 
              "sc_Macrophage_MARCO" , "sc_Cytotoxicity"  ,"sc_ExhCD8Tcell"  , "sc_Macrophage_MSR1"  , 
              "sc_Macrophage" ,"hallmark_HALLMARK_TGF_BETA_SIGNALING",
               "hallmark_HALLMARK_INTERFERON_ALPHA_RESPONSE","hallmark_HALLMARK_INTERFERON_GAMMA_RESPONSE", 
              "hallmark_HALLMARK_INFLAMMATORY_RESPONSE" ,"hallmark_HALLMARK_ANGIOGENESIS"  ,  "hallmark_HALLMARK_APOPTOSIS"          )


### Split criteria for the three categories
splitcat=Design$CD8IMMPH
names(splitcat)=rownames(Design)

### Choose order: based on signature e.g. Teffector 
sortSig="hallmark_HALLMARK_INTERFERON_GAMMA_RESPONSE"

### Reorder matrix based on sortSig & splitcat 
ematrixPlot=sig_scores[choicesig,names(sort(sig_scores[sortSig,],decreasing=T))]
catOrder=unlist(sapply(names(c(table(splitcat))),function(x) colnames(ematrixPlot)[which(colnames(ematrixPlot)%in%names(splitcat[which(splitcat%in%x)]))]))
ematrixPlot=ematrixPlot[,catOrder]

tmp=scaleRows(ematrixPlot)
rownames(tmp)=gsub("hallmark_HALLMARK_","H_",rownames(tmp))
###Plot the heatmap
svg("../figures/fig2a_signature_heatmap_HA.svg",width=15,height=6)
Heatmap(tmp,name = "rowZscore",
        cluster_columns=FALSE,
        use_raster=FALSE, 
        row_names_gp = gpar(fontsize = 12),column_names_gp = gpar(fontsize = 0.1),
        top_annotation=HeatmapAnnotation(IHC=splitcat[catOrder],
        Indication=Design[catOrder,"INDICAT"]))
dev.off()

```

## Signature boxplots per IHC  - full data  {.tabset}

```{r signatures bpxplot , results='asis', fig.height = 4, fig.width = 4}

idx1 = choicesig
my_comparisons <- list( c("INFLAMED", "EXCLUDED"), c("INFLAMED", "DESERT"), c("EXCLUDED", "DESERT"))


for(signame in idx1) {

  toplot=data.frame(score=as.numeric(sig_scores[signame,]),IHC_Status=as.character(Design$CD8IMMPH))
  g <- ggplot(data=toplot,aes(x = IHC_Status, y=score, fill = IHC_Status)) +
  geom_boxplot(outlier.shape = NA) + geom_jitter(cex = 0.5) + xlab("") + ylab(signame) + 
    stat_compare_means(comparisons = my_comparisons, method = "wilcox.test", p.adjust.method = "none", size=4, label = "p.signif") + theme_bw()
  
  cat("###", as.character(signame), "\n")
  #svg(paste0("../figures/fig2b_signature_boxplots_",signame,".svg"),width=6,height=7)
  print(g)
  #dev.off()
  cat("\n\n")
  
  
}


```

# HVG analysis

## PCA and correlation

```{r pca_percent_var, fig.height=2.5, fig.width=3.5, results="asis"}



includePCA <- TRUE
minNumSamplesPCA <- 3
topn=1500
  pander::pandoc.header("Percent variance explained {.tabset -}", level = 3)

  # Plot percent variance explained by top 10 PCs
  rv <- matrixStats::rowVars(dgelog$counts)
    select <- order(rv, decreasing = TRUE)[seq_len(min(topn,  length(rv)))]
  pca <- stats::prcomp(t(dgelog$counts[select, ]))
  percentVar <- pca$sdev^2/sum(pca$sdev^2)
  pca <- list(prcomp = pca, percentVar = percentVar)
    
  print(plotPercentVar(pca$percentVar, numPCs = 10))

```

```{r pca_plots, fig.width=8, fig.height=5, results="asis"}

  includeTissueSpecificPlots=FALSE
  pca_plot_x=c(1, 1, 3)
  pca_plot_y=c(2, 3, 4)
  
    
  pander::pandoc.header("PCA {.tabset -}", level = 3)
  endifnot(ncol(dge$counts) > 2, msg = "PCA analysis with less than 3 samples makes no sense")
  
  # PCs to plot
  dimensions <- data.table(x = pca_plot_x, y = pca_plot_y)
  # Keep only pairs with valid PCs (there are at most N-1 PCs, where N is the number of samples)
  dimensions <- dimensions[x < ncol(dge$counts) & y < ncol(dge$counts)]
  dimensions[, x := paste0("PC", x)]
  dimensions[, y := paste0("PC", y)]
  

  # Plot specified PCs
  headingLevel=4
  dt <- data.table(cbind(pca$prcomp$x, Design))
  pV <- round(pca$percentVar * 100)
  names(pV) <- paste0("PC", 1:length(pV))
  for (i in 1:nrow(dimensions)) {
        pander::pandoc.header(paste(paste(dimensions[i, ], collapse = "-"), 
            " {.tabset -}"), level = 4)
        xlab <- paste0(dimensions[i, "x"], " (", pV[dimensions$x[i]], 
            "% Variance Explained)")
        ylab <- paste0(dimensions[i, "y"], " (", pV[dimensions$y[i]], 
            "% Variance Explained)")
        labelVar = "SampleName"
        for (colorVar_k in colorVars) {
            pander::pandoc.header(paste(colorVar_k, " {-}"), 
                level = headingLevel + 1)
            labelVarBkp <- labelVar
            if (is.null(labelVar) & startsWith(colorVar_k, "QC")) {
                dt[, `:=`(Label, NA_character_)]
                dt[get(colorVar_k) != "", `:=`(Label, SampleName)]
                labelVar <- "Label"
            }
            gg <- ggplotPca(dt, x = dimensions[i, x], y = dimensions[i, 
                y], shape = NULL, colour = colorVar_k,
                 xlab = xlab, 
                ylab = ylab)
            print(gg)
            pander::pandoc.p("")
            labelVar <- labelVarBkp
        }
    }

```


$R^2$ for each single explanatory variable is shown against each dependent Principal Components variable separately. 
 
```{r var_cor_rsq, results="asis"}

  var_cor_num_PCs=15
  var_cor_rsq_threshold_pct=25
  
  pander::pandoc.header("Top variables correlated with PCs {.tabset -}", level = 3)
  
  # Step1: Identifies the features correlated to the top numPCs principal components
  # Number of principal components to include
  numPCs <- var_cor_num_PCs
  if (numPCs >= ncol(eset)) {
    numPCs <- ncol(eset) - 1
  }
  # Usually no need to change
  topngenes <- topn
  # Produces the pca and calculates the ssr with the phenotype variables in the dge object.
  cf <- cf.analysis(object = dge, pcs.use = numPCs, rsq.threshold.pct = var_cor_rsq_threshold_pct, topn = topngenes)
  
  # Used to set fig sizes in chunks below
  numTopVariables <- nrow(cf$top.explanatory.variables)
  nrowRsqMat <- nrow(cf$rsq.mat)


```

```{r var_cor_rsq_plot, fig.height=2+0.26*nrowRsqMat, fig.width=9, results="asis"}
# Separate chunk so fig size can be set using rsq.mat
  # Plot all features vs the top "pcs.use" principal components 
  NMF::aheatmap(as.matrix(cf$rsq.mat[, 1:numPCs]*100), fontsize = 8.5, 
                txt = round(as.matrix(cf$rsq.mat[, 1:numPCs]*100), digits = 1), color=c("white","red"), 
                main = "R-squared values * 100 (%) of variables concerning relevant PCs")
  

```

## UMAP Plots - top  var genes

```{r umap, results="asis"}

pander::pandoc.header("UMAP results {.tabset -}", level = 3)

# Now perform UMAP on the normalized data
umap_results <- umap::umap(t(dgelog$counts[select, ]))

Design$ID=rownames(Design)

umap_plot_df <- data.frame(umap_results$layout) %>%
  # Turn sample IDs stored as row names into a column
  tibble::rownames_to_column("ID") %>%
  # Add the metadata into this data frame; match by sample IDs
  dplyr::inner_join(Design, by = "ID")

tocolor=c("INDICAT","CD8IMMPH","BMFLOC", "STATUS","BATCH","STUDYID")

for (i in 1:length(tocolor)) {
        pander::pandoc.header(paste(tocolor[i], " {.tabset -}"), level = 4)
  
  # Plot using `ggplot()` function and save to an object
  tmp=umap_plot_df
  colnames(tmp)[1:3]=c("ID","UMAP1","UMAP2")
  tmp$GROUP=umap_plot_df[,tocolor[i]]
  final_annotated_umap_plot <- ggplot(
    tmp,
    aes(
      x = UMAP1,
      y = UMAP2,
      # plot points with different colors for each `refinebio_treatment` group
      color = GROUP,
      # plot points with different shapes for each `refinebio_disease` group
      shape = assay_type
    )
  ) +
    geom_point() + theme_bw() # make a scatterplot

  # Display the plot that we saved above
  print(final_annotated_umap_plot)
  pander::pandoc.p("")
  
}

#pca$prcomp$x

```
# Summary Model genes analysis: 0.05



## UMAP Plots - summary model genes: 0.05

```{r umap model genes, results="asis"}

pander::pandoc.header("UMAP results model genes {.tabset -}", level = 3)

features_final <- aws.s3::s3read_using(FUN = read.csv,object ="features_final_model.tsv", bucket = params$output_collection, sep="\t")
features_summary <- aws.s3::s3read_using(FUN = read.csv,object ="feature_summary.tsv", bucket = params$output_collection, sep="\t")
features_final=features_final[which(!features_final$feature%in%"(Intercept)"),]
features_summary=features_summary[features_summary$prop>0.05,]
features_final_all=features_final

features_final=features_final[features_final$feature%in%features_summary[features_summary$prop>=0.85,]$feature,]

rownames(ematrix)=make.names(rownames(ematrix))
# Now perform UMAP on the normalized data
umap_results <- umap::umap(t(ematrix[features_summary$feature, ]))

Design$ID=rownames(Design)

umap_plot_df <- data.frame(umap_results$layout) %>%
  # Turn sample IDs stored as row names into a column
  tibble::rownames_to_column("ID") %>%
  # Add the metadata into this data frame; match by sample IDs
  dplyr::inner_join(Design, by = "ID")



for (i in 1:length(tocolor)) {
        pander::pandoc.header(paste(tocolor[i], " {.tabset -}"), level = 4)
  
  # Plot using `ggplot()` function and save to an object
  tmp=umap_plot_df
  colnames(tmp)[1:3]=c("ID","UMAP1","UMAP2")
  tmp$GROUP=umap_plot_df[,tocolor[i]]
  final_annotated_umap_plot <- ggplot(
    tmp,
    aes(
      x = UMAP1,
      y = UMAP2,
      # plot points with different colors for each `refinebio_treatment` group
      color = GROUP,
      # plot points with different shapes for each `refinebio_disease` group
      shape = assay_type
    )
  ) +
    geom_point() + theme_bw() # make a scatterplot

  # Display the plot that we saved above
  print(final_annotated_umap_plot)
  pander::pandoc.p("")
  
}

#pca$prcomp$x

```

## PCA and correlation

```{r pca_percent_var_model, fig.height=2.5, fig.width=3.5, results="asis"}


topn <- dim(features_summary)[1]
select=rownames(dge$genes[dge$genes$GeneSymbol%in%features_summary$feature,])

  pander::pandoc.header("Percent variance explained {.tabset -}", level = 3)

  # Plot percent variance explained by top 10 PCs
  pca <- stats::prcomp(t(dgelog$counts[select, ]))
  percentVar <- pca$sdev^2/sum(pca$sdev^2)
  pca <- list(prcomp = pca, percentVar = percentVar)
    
  print(plotPercentVar(pca$percentVar, numPCs = 10))

```

```{r pca_plots_model, fig.width=8, fig.height=7, results="asis"}


    
  pander::pandoc.header("PCA {.tabset -}", level = 3)
  endifnot(ncol(dge$counts) > 2, msg = "PCA analysis with less than 3 samples makes no sense")
  
  # PCs to plot
  dimensions <- data.table(x = pca_plot_x, y = pca_plot_y)
  # Keep only pairs with valid PCs (there are at most N-1 PCs, where N is the number of samples)
  dimensions <- dimensions[x < ncol(dge$counts) & y < ncol(dge$counts)]
  dimensions[, x := paste0("PC", x)]
  dimensions[, y := paste0("PC", y)]
  

  # Plot specified PCs
  headingLevel=4
  dt <- data.table(cbind(pca$prcomp$x, Design))
  pV <- round(pca$percentVar * 100)
  names(pV) <- paste0("PC", 1:length(pV))
  
  for (i in 1:nrow(dimensions)) {
        pander::pandoc.header(paste(paste(dimensions[i, ], collapse = "-"), 
            " {.tabset -}"), level = 4)
        xlab <- paste0(dimensions[i, "x"], " (", pV[dimensions$x[i]], 
            "% Variance Explained)")
        ylab <- paste0(dimensions[i, "y"], " (", pV[dimensions$y[i]], 
            "% Variance Explained)")
        labelVar = "SampleName"
        for (colorVar_k in colorVars) {
            pander::pandoc.header(paste(colorVar_k, " {-}"), 
                level = headingLevel + 1)
            labelVarBkp <- labelVar
            if (is.null(labelVar) & startsWith(colorVar_k, "QC")) {
                dt[, `:=`(Label, NA_character_)]
                dt[get(colorVar_k) != "", `:=`(Label, SampleName)]
                labelVar <- "Label"
            }
            svg(paste0("../figures/sup_pca_plot_",colorVar_k,"_.svg"),width=8,height=6)
            gg <- ggplotPca(dt, x = dimensions[i, x], y = dimensions[i, 
                y], shape = NULL, colour = colorVar_k,
                 xlab = xlab, 
                ylab = ylab)
            print(gg)
            dev.off()
            pander::pandoc.p("")
            labelVar <- labelVarBkp
        }
    }

```

$R^2$ for each single explanatory variable is shown against each dependent Principal Components variable separately. 
 
```{r var_cor_rsq_modelg, results="asis"}


  pander::pandoc.header("Top variables correlated with PCs {.tabset -}", level = 3)
  
  # Usually no need to change
  topngenes <- dim(features_summary)[1]
  # Produces the pca and calculates the ssr with the phenotype variables in the dge object.
  cf <- cf.analysis(object =  dge[dge$genes$GeneSymbol%in%features_summary$feature,],
                    pcs.use = numPCs, rsq.threshold.pct = var_cor_rsq_threshold_pct, topn = topngenes)
  
  # Used to set fig sizes in chunks below
  numTopVariables <- nrow(cf$top.explanatory.variables)
  nrowRsqMat <- nrow(cf$rsq.mat)


```


```{r var_cor_rsq_plot_modelg, fig.height=2+0.26*nrowRsqMat, fig.width=9, results="asis"}
# Separate chunk so fig size can be set using rsq.mat
  # Plot all features vs the top "pcs.use" principal components 
  NMF::aheatmap(as.matrix(cf$rsq.mat[, 1:numPCs]*100), fontsize = 8.5, 
                txt = round(as.matrix(cf$rsq.mat[, 1:numPCs]*100), digits = 1), color=c("white","red"), 
                main = "R-squared values * 100 (%) of variables concerning relevant PCs")
  

```

# Final model gene analysis: 0.85

## UMAP Plots - final model genes: 0.85

```{r umap model genes final, results="asis"}

pander::pandoc.header("UMAP results model genes {.tabset -}", level = 3)

# Now perform UMAP on the normalized data
umap_results <- umap::umap(t(ematrix[features_final$feature, ]))

Design$ID=rownames(Design)

umap_plot_df <- data.frame(umap_results$layout) %>%
  # Turn sample IDs stored as row names into a column
  tibble::rownames_to_column("ID") %>%
  # Add the metadata into this data frame; match by sample IDs
  dplyr::inner_join(Design, by = "ID")



for (i in 1:length(tocolor)) {
        pander::pandoc.header(paste(tocolor[i], " {.tabset -}"), level = 4)
  
  # Plot using `ggplot()` function and save to an object
  tmp=umap_plot_df
  colnames(tmp)[1:3]=c("ID","UMAP1","UMAP2")
  tmp$GROUP=umap_plot_df[,tocolor[i]]
  final_annotated_umap_plot <- ggplot(
    tmp,
    aes(
      x = UMAP1,
      y = UMAP2,
      # plot points with different colors for each `refinebio_treatment` group
      color = GROUP,
      # plot points with different shapes for each `refinebio_disease` group
      shape = assay_type
    )
  ) +
    geom_point() + theme_bw() # make a scatterplot

  # Display the plot that we saved above
  #svg(paste0("../figures/sup_umap_plot_",tocolor[i],"_.svg"),width=8,height=6)
  print(final_annotated_umap_plot)
  #dev.off()
  pander::pandoc.p("")
  
}

#pca$prcomp$x

```

# GENE ANALYSIS

## Gene Heatmap Full

```{r gene_full, warning=FALSE, comment=FALSE, fig.height = 14, fig.width = 14, fig.align = "center"}


choicesig=unique(features_final_all$feature)
coefplot=features_final_all[which(!duplicated(features_final_all$feature)),]$coef
names(coefplot)=choicesig
coefplot=abs(coefplot)
#coefplotneg=coefplot; coefplotneg[coefplot>0]=0; #coefplotneg[coefplotneg<-0.5]=-0.5
#coefplot[coefplot<0]=0; #coefplot[coefplot>0.5]=0.5

### Split criteria for the three categories
splitcat=Design$CD8IMMPH
names(splitcat)=rownames(Design)

### Choose order: based on signature e.g. Teffector 
sortSig="CXCL9"

### Reorder matrix based on sortSig & splitcat 
ematrixPlot=ematrix[choicesig,names(sort(ematrix[sortSig,],decreasing=T))]
catOrder=unlist(sapply(names(c(table(splitcat))),function(x) colnames(ematrixPlot)[which(colnames(ematrixPlot)%in%names(splitcat[which(splitcat%in%x)]))]))
ematrixPlot=ematrixPlot[,catOrder]

tmp=scaleRows(ematrixPlot)
###Plot the heatmap
svg("../figures/sup5a_gene_heatmap_HA.svg",width=15,height=10)
Heatmap(tmp,name = "rowZscore",
        cluster_columns=FALSE,
        use_raster=FALSE, 
        row_names_gp = gpar(fontsize = 6),column_names_gp = gpar(fontsize = 0.1),
     left_annotation=rowAnnotation(abscoef=coefplot),
        top_annotation=HeatmapAnnotation(IHC=splitcat[catOrder],
        Indication=Design[catOrder,"INDICAT"]))
dev.off()
       #
```

## Gene Heatmap selected

```{r gene, warning=FALSE, comment=FALSE, fig.height = 5, fig.width = 14, fig.align = "center"}


choicesig=unique(features_final$feature)
coefplot=coefplot[choicesig]
#coefplotneg=coefplotneg[choicesig]


### Split criteria for the three categories
splitcat=Design$CD8IMMPH
names(splitcat)=rownames(Design)

### Choose order: based on signature e.g. Teffector 
sortSig="CXCL9"

### Reorder matrix based on sortSig & splitcat 
ematrixPlot=ematrix[choicesig,names(sort(ematrix[sortSig,],decreasing=T))]
catOrder=unlist(sapply(names(c(table(splitcat))),function(x) colnames(ematrixPlot)[which(colnames(ematrixPlot)%in%names(splitcat[which(splitcat%in%x)]))]))
ematrixPlot=ematrixPlot[,catOrder]

tmp=scaleRows(ematrixPlot)
###Plot the heatmap
svg("../figures/fig5a_gene_heatmap_HA.svg",width=15,height=6)
Heatmap(tmp,name = "rowZscore",
        cluster_columns=FALSE,
        use_raster=FALSE, 
        row_names_gp = gpar(fontsize = 12),column_names_gp = gpar(fontsize = 0.1),
        left_annotation=rowAnnotation(abscoef=coefplot),
        top_annotation=HeatmapAnnotation(IHC=splitcat[catOrder],
        Indication=Design[catOrder,"INDICAT"]))
dev.off()

```

## Genes boxplots per IHC  - full data  {.tabset}

```{r genes bpxplot , results='asis', fig.height = 4, fig.width = 4}

idx1 = unique(features_final$feature)
my_comparisons <- list( c("INFLAMED", "EXCLUDED"), c("INFLAMED", "DESERT"), c("EXCLUDED", "DESERT"))

for(signame in idx1) {

  toplot=data.frame(score=as.numeric(ematrix[signame,]),IHC_Status=as.character(Design$CD8IMMPH))
  g <- ggplot(data=toplot,aes(x = IHC_Status, y=score, fill = IHC_Status)) +
  geom_boxplot(outlier.shape = NA) + geom_jitter(cex = 0.5) + xlab("") + ylab(signame) + 
    stat_compare_means(comparisons = my_comparisons, method = "wilcox.test", p.adjust.method = "none", size=4, label = "p.signif") + theme_bw()
  
  cat("###", as.character(signame), "\n")
  #svg(paste0("../figures/fig5csupp5_gene_heatmap_",signame,".svg"),width=6,height=7)
  print(g)
  #dev.off()
  cat("\n\n")
}

```
